some code improvement
